requirements.txt
python-mpd2
Flask
PyYAML
